pyproject.toml
[project]
name = "thumbanana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.3.0",
]
